core-forecast-baseline/core-forecast-baseline-step-function/core_forecast_baseline_step_function/requirements.txt
pyYAML==5.4
cfn-flip